// src/app/core/models/technologie.model.ts
export class Technologie {
    idTechnologie?: number;
    titreTechnologie?: string;
    nomTechnologie?: string;
    description?: string;
    auteurTechnologie?: string;
    lienDocumentation?: string;
    typeTechnologie?: string;
    categorieTechnologie?: string;
    horodatageCreation?: Date;
}